refactor(integration-tests-lro): move tests for fake
Keep the tests closer to the thing they are testing and the top-level
module tidy.
